QA10-044 use Entity_Ids for frontend cross-references
With frontend cross-references directly picked from the memory (and not from
an ALI file), there is no need to convert them to and from strings.

This patch mostly just propagates the change in representation of scopes and
objects in the SPARK xrefs tables: from String_Ptr to Entity_Id.

Generated globals, phase 1, now represents references to __HEAP as a special
entity (which is the same representation as in the frontend xrefs), not as
an empty entity.

No detailed changelog entry due to size and mostly type-level effect of this
patch. No (intentional) changes of the behaviour.

Change-Id: I395a5c87d7752a501bad2d5f615e62a07142f5dd