QA17-021 interrupt handlers are similar to tasks
Interrupt handlers require exactly the same checks as tasks for accesses to
unsynchronized global variables. The easiest way to implement such checks is
to pretend that each protected object with an interrupt handler is a task.

The simplicity of this approach seems to outweigh two issues:

1) in error messages interrupt handlers appear as "tasks"

2) for interrupt handlers that shares a suspension object or an entry with
   other task, we will get spurious error about the shared ownership;
   however, even without such a shared ownership those interrupt handlers
   will be flagged as calling potentially blocking operations.

* gnat2why/flow/flow-control_flow_graph.adb
(Find_Tasks): detect interrupt handlers as they would be tasks objects.

* testsuite/gnatprove/tests/QA17-021__flow_handler/test.out
Update output; new checks.

Change-Id: Ic2b9c7437bad0f09d47f6f36e061f1c5b8e176fc