PC16-032 fix crash for very large record aggregates
This code in why-gen-records was generating a deep tree of
conjunctions. Reorganize to create a flat list of conjuncts instead.

The use of New_And_Then_Expr instead of New_And_Expr was a mistake; in
predicates (note that EW_Pred domain is hardcoded here), both are the
same.  Using New_And_Expr is not only clearer, it allows us to use the
flat list form.

We do this only for the record fields, out of convenience: it makes it
easier to compute the number of cells for the temporary conjuncts
array. Also, only the fields processing needs this, we can assume that
code doesn't have a large number of discriminants.

* why-gen-records.adb
(New_Field_Equality): rewrite code to use list form of conjunction
  instead of deep tree

Change-Id: I666c2e663d95ccdab79dfe1e540caf9df39300ae