QA10-044 rewrite retrieving of cross-references from the frontend
This patch finishes re-architecturing of how cross-references are retrieved
from the frontend.

Instead of sorting & removing duplicates in the frontend (where this is hard
because of language restrictions), we now traverse both cross-references and
cross-dereferences in the backend (where standard Ada sets are allowed).
This way sorting becomes unnecessary and removal of duplicates becomes
trivial. Also, we no longer create an intermediate copy of frontend xrefs in
an extra table, which saves us some memory.

No detailed changelog entry because the patch is rather big and little
remained from the old code. Behaviour remains as it was.

Change-Id: I55eb7cb3deccdf3c731d801ddd6edda32ea47de6