Q503-007 Fix memory leak in ce
This commit solves a memory leak in ce where we would initialize an
Other_Ptr variable with a value and then overwrite it with an other
allocated object. We now initialize its value to null then allocate it
in a loop. If no allocation is done, we set its value to Invalid.
Now Get_Typed_Cntexmp_Value is a function not a procedure.

* common/vc_kinds.adb
(Get_Typed_Cntexmp_Value) : Procedure to function change.

* common/vc_kinds.ads
(Cntexmp_Value_Ptr): now an access constant.
(Float_Value_Ptr): now a non null access constant.

Change-Id: I662664598ba1e830a4ac466250690ef911e1082b