QA10-044 remove special case for the dual-entity __HEAP variable
The special variable __HEAP was represented either by a special
SPARK_Xrefs.Heap entity (when it was coming from the frontend xrefs) or by
an Empty entity (when it was coming from a failed conversion from
Entity_Name).

This special treatment is no longer needed, because the backend retrieves
frontend cross-references already as Entity_Ids and no conversion is needed.

* gnat2why/utils/common_containers.adb
(To_Name_Set): remove special-case for the HEAP entity.

Change-Id: Ifd008ae60664b8677cb68298b84e800238e3acdf