LHS and RHS of Initializes contracts are disjoint
Both frontend and flow analysis are quite careful to only allow package
state variables on the LHS (outputs) and external variables on the
RHS (inputs) of an Initializes contract. There is no need to check for
actual inputs appearing as the contract outputs.

(no-tn-check)

* gnat2why/flow/flow-analysis.adb
(Check_Initializes_Contract): remove dead condition for actual inputs
 appearing on the LHS of Initializes contracts.

Change-Id: I821a90f68372c14053990951418d879e216b039f