Q925-059 fix assumption in loop
There was a too strict test for an expression to be "Void", which resulted
in a hypothesis not being generated. We introduce an extra function with a
more relaxed test.

* why-gen-expr.adb
(Is_Essentially_Void): new function to test if an EW_Prog_ID is void, but
  allows labels as well
* gnat2why-expr-loops.adb
(Transform_Loop_Statement): use Is_Essentially_Void instead of incorrect
  test

Change-Id: I75159805d734054698dbae599d3967d5a185b5b1