Q831-012 gprconfig wrapper
A new wrapper that prevents repeated invocations of gprconfig when running
the spark2014 testsuite. Such invocations are quite slow, because gprconfig
parses its XML configuration files.

This wrapper might cause setup problems, so its executable gets a very
descriptive name (to avoid conflicts) and is meant to be explicitly renamed
and copied to PATH (to actually make it work).

* gnatprove/memcache_client.ads
* gnatprove/memcache_client.adb
Move to common/, because they will be shared with the wrapper.

* scripts/gprconfig_memcache_wrapper/
Add wrapper.

Change-Id: Iba50769adb2e3841acb9f5cfc04e1ed9be2ba5b7