P926-007 fix handling of nested abstract states
With this patch, the vertices for abstract states that are constituents of
other states are added when processing their declarations (where they need
to have a Part_Of indicator), not when processing the Refined_State contract
where they appear as constituents.

An exception is made for abstract states declared in private child packages:
their Part_Of declarations are not processed, so they must be picked when
processing the Refined_State.

Also, add ??? comments to indicate further problems, as this commit is only
to fix a regression.

* gnat2why/flow/flow-control_flow_graph.adb
(Create): only process constituents declared outside of the main unit, i.e.
 in child packages.

(Do_Package_Declaration): don't expect vertices for abstract states of
 nested packages to be introduced elsewhere.

* testsuite/gnatprove/tests/N814-006__flow_spurious_ineffective_init/
* testsuite/gnatprove/tests/N814-007__refined_post/
Remove warnings about ineffective statements (regression) and regression
comments in the tests; new info messages about abstract state that are
constituents of states that are listed in the Initializes aspect. The new
messages appear because vertices are now added in Do_Package_Declaration and
their Is_Export attribute is set; previously they were added in Create,
where setting of the Is_Export attribute was (most likely) forgotten.

* testsuite/gnatprove/tests/M729-011__flow_body_stubs/test.out
* testsuite/gnatprove/tests/MB26-023__flow_nested_packages_effects/test.out
* testsuite/gnatprove/tests/O909-059__part_of/test.out
New info messages (for the same reason as explained above).

Change-Id: I1c572fe069a86c2d0d816eed1fcb107d858482a8