Q929-028 relax precondition
* flow_types.ads
(Flow_Id_To_String): Revert a part of the commit for Q929-028 - the
precondition for Flow_Id_To_String was only ever meant to check the
essential, not validate if the graph constructed is right. Right now it
causes gnatprove to crash in debug mode on for example Q710-004__range.

Change-Id: Ie6998caef25858f8aaae94317cd9c2a6daeb3933