QB01-007 remove distinction between ghost and non-ghost Initializes
Data read during package elaboration can flow to both ghost and non-ghost
package state variables, i.e. we cannot make distinction based only on the
"ghostness" of the item that appears on the LHS of the Initializes contract.

To narrow the flows captured by the generated Initializes contract we would
need to know the "ghostness" of the item that appears on the RHS:

   LHS ghost | RHS ghost | Allowed?
   --------------------------------
      No     |   No      |  Yes
      No     |   Yes     |  No  (*)
      Yes    |   No      |  Yes
      Yes    |   Yes     |  Yes

(*) Actually, even this can be allowed when a ghost constituent of a
non-ghost abstract state is initialized with a ghost input.

However, we currently don't store this info in the ALI file (this is the
subject of Q515-016). Perhaps we should, but then we need a test where this
makes a difference (and apparently we don't have such a test yet). And then,
we don't need to store package state as ghost and non-ghost variables,
because their ghost status will be provided by Q515-016.

I suspect the distinction between ghost and non-ghost own variables was
motivated by the same misinterpretation of the SPARK RM that caused reads in
ghost code to be classified as Proof_In. (Similarly, it was wrong to include
Proof_Ins of subprograms called during package elaboration on the RHS of the
generated Initializes contract, which AFAIU is what the old GG did).

Affected code includes package state discovery (phase 1), serialization
(phase 1 and 2) and final generation of the Initializes contract (phase 2).

No impact on the testsuite.

Change-Id: Ibae7e0ce6f1c8dc7f385b5054174a450d7d88a61