QA25-021 remove AS_In_Generated_Initializes
Routine AS_In_Generated_Initializes was introduced in P330-009, where we
dealt with a wrongly diagnosed issue. This patch reverts it to make the code
cleaner before finding a real solution to problems of that and this TN.

* gnat2why/flow/flow-analysis.adb
(AS_In_Generated_Initializes): remove.
(Find_Use_Of_Uninitialized_Variables): don't suppress messages based on
 AS_In_Generated_Initializes.

* testsuite/gnatprove/tests/TU__tasks_legal/test.out
Remove info messages about initialization of a protected object before calls
to its protected operation; this is consistent with not-emitting similar
messages for P_Int (which unlike Hidden_PO is declared in the visible part
of a package spec and is not a constituent of any abstract state).

Change-Id: Ibde36feb4876401dfb064482722d516339d9edd3