Q912-011 Avoid printing of out of range index for ce
This commits treat a case which happens when index of arrays of the model
are given outside of their enumeration type range by cvc4.
This commit comes with the modification of Get_Enum_Lit_From_Pos in
sem_eval.adb.

* gnat2why/why/gnat2why-counter_examples.adb
(Refine_aux): Becomes Refine_aux to avoid confusion with Refine. Add a
boolean parameter: if true, refine is refining an index. A case is added
when an error is raised by Get_Enum_Lit_From_Pos so that nothing is
printed for such a counterexample value.

Change-Id: I599d1da9a91a03556f64d642beaaae9a328c35a8