QB07-050 reject references to single concurrent types
While marking default expressions in single concurrent objects the current
concurrent type is denoted by an E_Variable (because that is more convenient
for the frontend) and must be converted to a type entity (because that is
that the gnat2why generally expects to see).

We already had a Canonical_Entity routine for dealing with this while
parsing flow contracts; now it moved, adapted and reused in marking.

* gnat2why/flow/flow_utility.ads
* gnat2why/flow/flow_utility.adb
(Canonical_Entity): move to SPARK_Util.

* gnat2why/spark/spark_util.ads
(Canonical_Entity): allow calls from Context of an E_Protected_Type and
 record types; update comment.

* gnat2why/spark/spark_definition.adb
(Is_Current_Instance): replace Unique_Entity with Canonical_Entity; rename
 E to Ref to remove shadowing.

* testsuite/gnatprove/tests/QB07-050__single_po/test.out
Update output; new error emitted from marking.

Change-Id: I97523b8640263edf358ff3b9a4c7193756d41169