P118-034 change expected output of GG for private child package
To deduce that Refined_Global of F_6 is G we would have to see the
Refined_Global of its callee, F_4, which means that a child package would
have to see the parent's body.

Instead, the Refined_Global of F_6 is the same as the abstract global of its
callee, F_4, i.e. A.State_A.

* testsuite/gnatprove/tests/P118-034__flow_gg_private/test.out
Change expected output.

Change-Id: I86c9465737464ab4c7aa74fba859a25fedb997d4