QB03-022 Return with non-zero exit status when checks not proved
When checks are not proved, GNATprove now returns with a non-zero
exit status (of value 1). This makes it easier to stop analysis in
a Makefile-like integration when some checks are both not proved
and not justified.

Note that this does not apply to warnings, which continue to be considered
separately from errors/checks, unless --warnings=error is set, in which
case warnings are treated as errors, and lead to a non-zero exit status.

* common/call.ads
(Unproved_Checks_Error_Status): Special value for communication between
 spark_report and gnatprove.

* gnatprove/report_database.ads
(Has_Unproved_Check): Return whether there are unproved checks.

* gnatprove/spark_report.adb
(SPARK_Report): Return with a special exit status if there are unproved
 checks.

* gnatprove/gnatprove.adb
(Generate_SPARK_Report): When calling spark_report and it returns with
 the special exit status of Unproved_Checks_Error_Status, return an exit
 status of 1.

* docs/ug/en/source/how_to_run_gnatprove.rst
Update documentation of exit status.

* testsuite/gnatprove/lib/python/test_support.py
(gnatprove): Add parameter exit_status to check the expected exit status.

* testsuite/gnatprove/tests/QB03-022__exit_status/incr.adb
* testsuite/gnatprove/tests/QB03-022__exit_status/test.out
* testsuite/gnatprove/tests/QB03-022__exit_status/test.py
Add test.

Change-Id: I5336c1319323cf43a1c6e30ebf97a56aaa7d713c