QB06-003 fix unique entry ids in child packages
Routine Full_Entry_Name was meant return the Unique_Name of the entry call
target suffixed with names of the record components when the protected
object is nested in a composite object.

However, it only climbed the scope hierarchy up to Standard and for objects
declared in child packages this missed the name of the encapsulating parent.

This patch reimplements this Full_Entry_Name to traverse the prefix nodes
related to array and record components, and then resorting to much better
tested Unique_Name.

* gnat2why/spark/spark_util.ads
* gnat2why/spark/spark_util.adb
(Full_Entry_Name): reimplement.

* testsuite/gnatprove/tests/QB06-003__flow_child_entry/
Enable; add output.

Change-Id: Ia1a75afa0cc7891bec8bece5a51781ed4dd6a979