QB09-010 respect backend failures in modes check and check_all
The 'gnatprove' driver ignored 'gnat2why' failures with the intention to
"require compilable code for analysis" (quote from K621-029 where this was
implemented). But this was done years ago and apparently no longer worked.

With this patch gnatprove respects gnat2why failures and stops with either:

  gnatprove: error during generation of Global contracts

when the code is entirely invalid or

  gnatprove: error during fast partial checking of SPARK legality rules
                        | full checking of SPARK legality rules
                        | analysis of data and information flow
                        | flow analysis and proof

(depending on the --mode switch) if it contains SPARK violations.

In any case, when phase 1 terminated with a failure we should not go into
phase 2 in modes that rely on the ALI files being generated (i.e. check_all,
flow and prove), because we will crash due to missing synthesized contracts.
Apparently this patch does that.

* gnatprove/gnatprove.adb
(Execute_Step): do not suppress gnat2why failures in modes check and
 check_all.

* testsuite/gnatprove/tests/O429-045__tasking_restrictions/test.out
On SPARK violation in --mode=check we now exit with an error and not with a
summary table.

* testsuite/gnatprove/tests/P615-019__stone_level/test.out
Similar to the above, but this test calls gnatprove with "-q", which
prevented the message about summary table from appearing in test.out.

* testsuite/gnatprove/tests/QB09-010__check_all/
Enable and add output; no longer crashes.

Change-Id: Ia8edeb1204bd7219d4f1d64e156f7df1960b8408