QB10-014 add constituents from a private child to the parent's CFG
All constituents from the private children, i.e. abstract states, variables
and constants with variable inputs, are now added to the parent's CFG.

This patch merely adds the 'Initial and 'Final vertices to the CFG; what
remains to be done is the set the Is_Initialized attribute of the 'Initial
vertex, most likely based on the (generated) Initializes contract of the
private child.

* gnat2why/flow/flow-control_flow_graph.adb
(Create): add all constituents from the private children, not only abstract
 states.

* testsuite/gnatprove/tests/QB10-014__flow_priv_child/test.out
Check about X not being listed in the Initializes aspect is gone, but a new
spurious check about X being not initializes appeared. To be investigated
separately.

Change-Id: I6151fbc094e79e91af8fc4d76df42d805f95e472