exclude null pointers from the hash-consing package
Hash-consing never operates on null pointers. They are now excluded on a
type level. This makes the code more robust (because the compiler will not
allow to dereference a null pointer) and quite likely more
efficient (because the compiler is allowed to skip checks for pointers being
null).

(no-tn-check)

* common/hash_cons.ads
(Access_Type): now is a not-null access type.

* common/assumption_types.ads
(Subp_Type): propagate the type change from the Hash_Cons package.

* common/assumptions-search.adb
(Unverified_Claims): initialize Cur_Token at the point of declaration;
 required because its Subp_Type component being of a not-null type must be
 initialized to a non-null value.

Change-Id: Ibbf014f1ad0dfbd40213cb383604624dc373197d