Q608-016 Use safe assumption on array bounds for CodePeer in SPARK
CodePeer is used as prover in SPARK when --codepeer=on. It was making
an optimistic assumption that the low bound of array indexes is within
bounds of the index subtype, which is not always true for empty arrays.
For example, an empty string can range from -1 to -2, with a low bound
of -1 which is not a positive number.

The integration of CodePeer in SPARK now removes this unsound assumption
for proof.

* gnatprove/spark_codepeer_wrapper.adb
(Generate_SCIL): Use switch to force sound use of first bound in arrays.

* testsuite/gnatprove/tests/Q608-016__codepeer_string/test.opt
* testsuite/gnatprove/tests/Q608-016__codepeer_string/test.out
File correct output.

Change-Id: I416c7d3315705cb4f625fae7a785493fd0f8da0d