QB16-044 reenable Check_Generated_Refined_Global for expr funcs
Sanity routine Check_Generated_Refined_Global was most likely disabled for
expression functions to avoid spurious errors when such functions declared
in the visible part of the package spec were analysed in the scope of that
very part.

Now we analyse such functions in the scope of the package body and there is
no reason to skip Check_Generated_Refined_Global for them.

(Arguably, suppressing this routine for all expression functions was too
drastic.)

* gnat2why/flow/flow-analysis-sanity.adb
(Check_Generated_Refined_Global): don't skip this check for expression
 functions.

* testsuite/gnatprove/tests/sparksmt__lexer/lexer.ads
* testsuite/gnatprove/tests/sparksmt__lexer_termination/lexer.ads
Fix Global contracts: objects used to derive the result of a ghost function
are Inputs (they would be Proof_Ins if used only in assertion expressions).

Change-Id: Iaad838198af383c4ca9edad2d20937ea1998952c