QB17-009 Use Get_Variables for expr with variable inputs
For type predicates and user defined equalities we were retrieving the
variables used through Get_Globals on the subprograms. This feels wrong
and we are now using Get_Variables on the defining expression as it is
done for all the other expressions in this code.

* gnat2why/flow/flow-analysis-sanity.adb
(Global_Reads): removed
(Check_Expression): use Get_Variables on the current Expression instead of
Global_Reads.

* testsuite/gnatprove/tests/QB17-009__flow_dyn_pred
Test output updated.

Change-Id: Ife355554af1b57c93a771e0d4a1d2076b1c5dc0d