reenable test that no longer crashes
The euclidian_division test was disabled due to no frontend cross-references
being provided for not-in-SPARK instances of generic packages.

Quite likely, this was fixed by switching frontend cross-references from
Entity_Names (which were not reliable for generics, because they were picked
very early in the frontend analysis) to Entity_Ids.

(no-tn-check)

* testsuite/gnatprove/tests/euclidian_division/
Enable; add counterexample, remove warning (in similar code flow also no
longer warns about such an unused assignment), and remove a check about an
access to the __HEAP object, which is no longer picked by the frontend
globals (there is no obvious access in this in the instantiated package, so
it must have been coming from some deep callee in the predefined unit).

Change-Id: I6103942ded339dcf086be8d59bda623828491356