reenable tests on runs without assertions
Some tests were disabled because we would crash in the GG-sanity-assertions,
but in fact the violated assertion should have little impact on the expected
outputs. It seems better to have those tests enabled, as otherwise we are
loosing track of which commits affects their outputs.

(no-tn-check)

* testsuite/gnatprove/tests/O916-043__crazyflie/
* testsuite/gnatprove/tests/O916-052__crazyflie/
* testsuite/gnatprove/tests/O917-119__crazyflie/
* testsuite/gnatprove/tests/P725-016__recursive_spec/
Re-enable; accept the current output as it is, as it looks reasonable.

Change-Id: I8afae24b3a8d5608896b0319e91db8d6531dabad