PA30-001 fix abuses of Part_Of objects from SPARK_Mode => Off code
In this test we referencing Part_Ofs protected objects in a way, that should
typically be rejected by the frontend. However, frontend accepts this code
because we (intentionally for other reasons) have SPARK_Mode => Off.

The violations are most likely just copy-paste mistakes, as the current code
is not even accepted by the frontend.

With this patch, the frontend will at least be able to compile this test and
will keep accepting it even when the SPARK_Mode annotations are changed from
Off to On.

* testsuite/gnatprove/tests/PA30-001__non_spark_comp/
Only reference Part_Of objects from their corresponding single protected
owners.

Change-Id: I49a8f7272a9df4162997b8c923d07146166e36f4