QA24-010 Unsynch globals for subp with body not in SPARK
For subprograms annotated with a global contract but with body not in
SPARK, when calling Preanalyze_Spec we were not saving the unsynchronized
globals. We now are.

* gnat2why/flow/flow_generated_globals-partial.adb
(Preanalyze_Spec): when looking for Unsynchronized_Globals we now return
the correct set.

* testsuite/gnatprove/tests/O429-044__synchronous_barriers
* testsuite/gnatprove/tests/O917-095__flow_tasking_main_program
* testsuite/gnatprove/tests/QA12-010__flow_race
Test output updated. We now detect data races for subprograms of which
we only have a spec (or there is a body but is hidden behind a SPARK_Mode
=> Off).

Change-Id: I08571abba26078ded6876b62813db8fa2c8d675c