P517-019 add test for variable inputs in type declarations
This test contains 5 type declarations, each with some kind of error related
to variable input: either in the default expression of a discriminant [see
SPARK RM 3.7(3))] or (an implicit) subtype declaration [see SPARK RM
3.2.2(1)].

So we should get 5 errors and indeed, we do, but some of those errors are
duplicated (for T1 and T3, in default expressions in type completions) while
others are missing (for T2 and T4).

* testsuite/gnatprove/tests/P517-019__flow_types/
Add; some errors are missing, while other are duplicated.

Change-Id: I515487f3a7bfc44259456986b5e3673988cba5dd