reject calls to Entire_Variable for Flow_Ids with Null_Value
Calls to Entire_Variable are meaningless if the argument is of a Null_Value
kind; now they are rejected in the contract.

(no-tn-check)

* gnat2why/flow/flow_types.ads
(Entire_Variable): reject Null_Value in Pre and simplify Post;

* gnat2why/flow/flow_types.adb
(Entire_Variable): raise an exception on Null_Value parameters.

* gnat2why/flow/flow-slice.adb
(Flow_Equivalent): add contract and simplify; this routine was never called
 with Null_Value and now this is checked in the Entire_Variable contract.

Change-Id: I05d6e5c3fdd031c89429424c52254655c4e162a5