QB28-011 Update Tokeneer code towards fully proving it
Tokeneer code needs to be updated/fixed to be provable with GNATprove.
Here is a first set of modifications. Mostly, I moved some types/state from
the body of packages to the private part of the package spec, where they
can be visible for the proof of client code. In those cases, the proof in
SPARK 2005 used proof functions which did that at the level of VCs.

These changes go from 234 unproved checks to 102.

* testsuite/gnatprove/tests/tokeneer/admin.adb
* testsuite/gnatprove/tests/tokeneer/admin.ads
(FinishOp): Remove spurious precondition, not in SPARK 2005 version.
Move query function bodies to the spec.

* testsuite/gnatprove/tests/tokeneer/admintoken.adb
* testsuite/gnatprove/tests/tokeneer/admintoken.ads
Move types/state/queries to the spec.

* testsuite/gnatprove/tests/tokeneer/auditlog.adb
* testsuite/gnatprove/tests/tokeneer/auditlog.ads
Turn the manual proof of a package invariant in SPARK 2005 into an
automatic proof of a type invariant, on the new global variable
LogFileState which bundles 4 parts of the state related by an invariant.
Update code and contrats as a consequence. One intrusive change is that
local subprograms need to take LogFileState as parameter, otherwise
they require that the type invariant is set on entry and exit.

* testsuite/gnatprove/tests/tokeneer/cert-attr-auth.adb
* testsuite/gnatprove/tests/tokeneer/cert-attr-auth.ads
Move queries to the spec.

* testsuite/gnatprove/tests/tokeneer/clock.adb
* testsuite/gnatprove/tests/tokeneer/clock.ads
Move types/state/queries to the spec.

* testsuite/gnatprove/tests/tokeneer/door.adb
* testsuite/gnatprove/tests/tokeneer/door.ads
(Poll): Use if-then instead of Boolean less-than for the implication
 between two conditions. Remove part of the postcondition that it not
 provable (was proved by review in SPARK 2005).

* testsuite/gnatprove/tests/tokeneer/enclave.adb
* testsuite/gnatprove/tests/tokeneer/enclave.ads
Change "and" into lazy "and then" in places where the rhs needs to be
evaluated in a context where the lhs is known to hold, typically for
preconditions of functions.
Move types/state/queries to the spec.

* testsuite/gnatprove/tests/tokeneer/enrolment.adb
Add pragma Annotate for justification that repeats manual justification
in PRV file in SPARK 2005 version.

* testsuite/gnatprove/tests/tokeneer/tismain.adb
Change "and" into lazy "and then" in places where the rhs needs to be
evaluated in a context where the lhs is known to hold, typically for
preconditions of functions.

* testsuite/gnatprove/tests/tokeneer/keystore.adb
* testsuite/gnatprove/tests/tokeneer/keystore.ads
Move types/state/queries to the spec.

* testsuite/gnatprove/tests/tokeneer/latch.adb
* testsuite/gnatprove/tests/tokeneer/latch.ads
Move types/state/queries to the spec.

* testsuite/gnatprove/tests/tokeneer/proof/sessions/*/why3session.xml
Update session files.

* testsuite/gnatprove/tests/tokeneer/test.adc
Allow use of Old on expressions not always evaluated.

* testsuite/gnatprove/tests/tokeneer/test.opt
No reason not to run on Darwin, given that it's only a replay now.

Change-Id: I4761d4eff4988b4358fe25001fb0544af2f63905