QA12-027 Default initialization cannot look in private part
In this patch we have mainly added a Flow_Scope parameter to the routines
that deal with default initialization, i.e. Is_Default_Initialized and
Default_Initialization. This is because we were wrongly looking into
private parts of private types when checking for default initialization and
now we prevent this. Some of the patch is about moving routines in order
not to have circular dependency between packages.

* gnat2why/flow/flow-analysis.adb
* gnat2why/flow/flow-control_flow_graph-utility.ads
* gnat2why/flow/flow-control_flow_graph-utility.adb
* gnat2why/flow/flow-control_flow_graph.adb
Add Flow_Scope parameter to calls to Is_Default_Initialized.
In flow-control_flow_graph I have also slightly cleaned the conditions to
issue an error message. There is no need to check for the type being
private or not having its full type in SPARK as these things are taken into
account inside Default_Initialization.

* gnat2why/flow/flow_refinement.ads
* gnat2why/flow/flow_refinement.adb
(Default_Initialization): moved here routine from spark_util-types. When
Explicit_Only is false, if the type has an inherited DIC, is private and it
is not derived then we return No_Default_Initialization.

* gnat2why/flow/flow_types.ads
Moved here type Default_Initialization_Kind from spark_util.

* gnat2why/flow/flow_utility-initialization.ads
* gnat2why/flow/flow_utility-initialization.adb
(Is_Default_Initialized): add Flow_Scope parameter to routine.

* gnat2why/spark/spark_definition.adb
* gnat2why/spark/spark_util.adb
* gnat2why/why/gnat2why-counter_examples.adb
* gnat2why/why/gnat2why-expr.adb
Add Flow_Scope parameter to calls to Default_Initialization.

* gnat2why/spark/spark_util-tpyes.ads
* gnat2why/spark/spark_util-tpyes.ads
(Default_Initialization): removed

* testsuite/gnatprove/tests/MA04-032__flow_textio_issue
* testsuite/gnatprove/tests/Q328-034__mut_discr
* testsuite/gnatprove/tests/evoting
Test output updated with new check about an object of type
Ada.Text_IO.File_Type not being initialized.

* testsuite/gnatprove/tests/OA30-010__current_task_in_entry_body
* testsuite/gnatprove/tests/P607-037__crash_on_computed_calls
Edit test to initialize the component of the protected type of type
Ada.Task_Identification.Task_Id to Null_Task_Id.

* testsuite/gnatprove/tests/Q204-001__linted
Edit test adding a Default_Initial_Condition where it was needed to make
a protected type fully initialized.

* testsuite/gnatprove/tests/KB29-037__typinv2
* testsuite/gnatprove/tests/N711-045__flow_default_init_cond
* testsuite/gnatprove/tests/N729-011__flow_passively_used_variables
* testsuite/gnatprove/tests/O402-017__tagged
* testsuite/gnatprove/tests/P718-001__ubstr_bad
* testsuite/gnatprove/tests/P718-001__ubstr_good
* testsuite/gnatprove/tests/P820-002__protected_out
* testsuite/gnatprove/tests/QA12-027__flow_DIC
* testsuite/gnatprove/tests/QA12-027__flow_uninitialized_constituent
Test output updated. We now correctly detect that the objects mentioned in
the new checks are not initialized as their type is private and we cannot
see inside the private part.

* testsuite/gnatprove/tests/MB28-040__full_default_initialization
* testsuite/gnatprove/tests/N404-009__box_in_aggregates
* testsuite/gnatprove/tests/O403-005__private_attributes
* testsuite/gnatprove/tests/P816-010__prefix_inline
* testsuite/gnatprove/tests/sparksmt__lexer
* testsuite/gnatprove/tests/sparksmt__lexer_termination
Edit test and add Default_Initial_Condition not to get an unwanted (and
irrelevant for the test) error message.

* testsuite/gnatprove/tests/P801-003__flow_type_invariants
Test output updated. Now we detect an unused variable in a routine. This
variable is of a private type without DIC.

Change-Id: I7adc26e7cd39ba63fac86afeae9949e95af75638