QC05-057 add test for aliasing and non-returning subprograms
For this test we should get 2 checks about aliasing (in Test1 and Test2),
while the check about call to nonreturning subprogram is spurious (it is
fine for nonreturning subprograms to call other nonreturning subprograms).

The warning about G not being modified is completely misleading, but
probably this is a side-effect of undetected aliasing.

* testsuite/gnatprove/tests/QC05-057__flow_aliasing/
Add test; output is not as expected.

Change-Id: I88433c2f293bc5ef93d4079c80a7e2b5051e55b6