QC07-017 detect constants with Part_Of indicator
Detect constants without variable input not only in the Refined_State (which
is only available when the package body is in SPARK), but also within the
Part_Of constituents (which is available when the package spec is in SPARK).

* gnat2why/flow/flow-analysis.adb
(Check_Refined_State_Contract): also detect constants with Part_Of
 indicators.

(Check_Initializes_Contract): only flag constants without variable input
 declared in the visible part of the analyzed package.

* testsuite/gnatprove/tests/QC07-017__flow_constant_part_of/test.out
New check (this is the primary goal of this TN, it comes from the rewrite of
Check_Refined_State_Contract).

* testsuite/gnatprove/tests/O414-015__flow_constants_in_contracts...
* testsuite/gnatprove/tests/O903-024__flow_constituent_var_input/test.out
* testsuite/gnatprove/tests/TU__types_and_subtypes_illegal/test.out
Remove checks that were duplicating those for Refined_State contract (this
is an effect of the change in Check_Initializes_Contract).

Change-Id: Id7d08263125750045a067dbf4d0b7a58cf2c453d