QA17-002 Solve lemma diffs
A proof using altergo was removed by b4351fad because altergo was able
to do the proof at that time. It is not possible anymore (but will be
in 1.30). So, this proof is now done in Coq as before. The proof has been
splitted into two asserts as we had difficulties putting two proofs on the
same line *and* reliably use generate_session.py on this.
A hack was necessary to make cpp actually replace one of the variable
for unknown reasons (the order symbols are used on the specific line may be
to blame).

* include/manual_proof.in
Adding the new Coq proof.

* include/proof/Coq/common/lemma_modular_div_then_mult_bounds_part*.prf
Line break before SPARK_MAX_UINT_PLUS1 variable so that cpp replaces it.

* include/proof/sessions/*/why3session.xml
Update sessions.

* include/spark-mod_arithmetic_lemmas.adb
Separating the postcondition proof into two asserts.

* testsuite/gnatprove/tests/P405-006__coq_lemma_library/test.out
Updating tests

Change-Id: I4656f0bb60fee93852c3030d1842e2033430a866

Change-Id: Id9e25d38def9d04fca4693b5a2cde8c3d21b2c0a