QB28-011 Complete automatic proof of Tokeneer
Tokeneer is now fully proved automatically, with appropriate
justifications for unprovable checks (39 in total), most of them for
the range of 'Image results. There remains 6 unproved spurious predicate
checks, which will be fixed under QC05-058.

* tests/tokeneer/admin.adb
* tests/tokeneer/admin.ads
(AllowedOp): Move body to private part of spec for use in proof of
 client code.

* tests/tokeneer/enclave.adb
(StartAdminOp): Use shortcut "and then" for precondition to be provable.

* tests/tokeneer/proof/sessions/configuration/why3session.xml
* tests/tokeneer/proof/sessions/doorapi/why3session.xml
* tests/tokeneer/proof/sessions/enclave/why3session.xml
* tests/tokeneer/proof/sessions/screen/why3session.xml
* tests/tokeneer/proof/sessions/tismain/why3session.xml
* tests/tokeneer/proof/sessions/updates/why3session.xml
Update session files.

* tests/tokeneer/test.out
Update output.

Change-Id: Ifa600e608d47f20db309306a01939ace42cc47df