QC11-014 add test for impossible to satisfy check
Checks insisting on listing X and Y in the Initializes aspect of P is
impossible to satisfy (at least not without violating other rules). This
code is different from the one in my opening TN (that one was created with
some local changes), but it still demonstrates the point.

* testsuite/gnatprove/tests/QC11-014__flow_initializes/
Add; high checks are spurious, just like similar checks that I am suggesting
to (at least temporarily) accept as part of this TN.

Change-Id: Ic04ed9dfaa2aa0abda4648f842196acc5be049f3