Q929-022 Revert previous commit
In the previous commit under this ticket we were considering every bounds
to have variable inputs but the ones coming from formal parameters. This
was wrong and we are now considering every bound to be without variable
inputs.

Note that Q929-022 remains enabled as it was xfailed for a wrong reason so
we have now just updated its output.

Change-Id: Iaef24687a8e65efd583394af7e43d3a9348a618d