QC10-002 fix marking of delayed type aspects
For calls to Mark_Entity on subprograms that implement delayed type
aspects (i.e. Predicate, Default_Initial_Condition and Type_Invariant)
the Current_SPARK_Pragma is set by the callers.

In particular, it must not be reset to the SPARK_Pragma of that subprogram
Entity: first, because such subprograms can't be explicitly annotated with a
SPARK_Mode; second, because their SPARK_Pragma is inherited from where their
declaration occurs and that depends on where their corresponding type is
frozen (which is arbitrarily decided by the frontend and unrelated to the
SPARK_Mode rules).

* gnat2why/spark/spark_definition.adb
(Mark_Entity): don't reset Current_SPARK_Pragma for subprograms that
 implement delayed type aspects.

* testsuite/gnatprove/tests/QC10-002__predicate/
Enable test; add output.

Change-Id: Iff73f41a453517dea099fc605f3d7d18edb21813