QC08-004 test for visibility among Entity_Names
In this test Main is reading an abstract state that is only known by
Entity_Name and its initialization status, which depends on the
initialization of a private child's state, must be resolved in phase 2.

This suggests that the algorithm which now relies on Is_Visible must either
be adapted to not use it, or Is_Visible must be adapted to reliably work on
Entity_Names.

Note that this test still represents a relatively easy scenario, as the
abstract states are initialized in elaboration of their enclosing packages.
In the extreme case they would be only initialized by explicit calls to some
initialization procedures, an in the worst case we would also need to
generate Global contracts for those procedures.

* testsuite/gnatprove/tests/QC08-004__flow_visible/
Add; spurious check about Parent.State not being initialized.

Change-Id: I3c1691045bfd027b47d2946abb8dfa7135f24b93